Major changes
Adding a Huffingman.java class, which hold the classes for creating the huffingman tree, as well as helper methods. Renamed the class holding the main method to Main.class, and added some other code to other classes as well. This is still based solely on Conner's lab, which he did give us permission to do, but I want to make some changes after everything works as I think there are better ways to do some things.